+++ src/oscar/apps/basket/forms.py
@@ -1,7 +1,6 @@
 from django import forms
 from django.conf import settings
 from django.db.models import Sum
-from django.forms.models import BaseModelFormSet, modelformset_factory
 from django.utils.translation import ugettext_lazy as _
 
 from oscar.core.loading import get_model
@@ -50,32 +49,6 @@ class BasketLineForm(forms.ModelForm):
         fields = ['quantity']
 
 
-class BaseBasketLineFormSet(BaseModelFormSet):
-
-    def __init__(self, strategy, *args, **kwargs):
-        self.strategy = strategy
-        super(BaseBasketLineFormSet, self).__init__(*args, **kwargs)
-
-    def _construct_form(self, i, **kwargs):
-        return super(BaseBasketLineFormSet, self)._construct_form(
-            i, strategy=self.strategy, **kwargs)
-
-    def _should_delete_form(self, form):
-        """
-        Quantity of zero is treated as if the user checked the DELETE checkbox,
-        which results in the basket line being deleted
-        """
-        if super(BaseBasketLineFormSet, self)._should_delete_form(form):
-            return True
-        if self.can_delete and 'quantity' in form.cleaned_data:
-            return form.cleaned_data['quantity'] == 0
-
-
-BasketLineFormSet = modelformset_factory(
-    Line, form=BasketLineForm, formset=BaseBasketLineFormSet, extra=0,
-    can_delete=True)
-
-
 class SavedLineForm(forms.ModelForm):
     move_to_basket = forms.BooleanField(initial=False, required=False,
                                         label=_('Move to Basket'))
@@ -107,23 +80,6 @@ class SavedLineForm(forms.ModelForm):
         if not is_available:
             raise forms.ValidationError(reason)
         return cleaned_data
-
-
-class BaseSavedLineFormSet(BaseModelFormSet):
-
-    def __init__(self, strategy, basket, *args, **kwargs):
-        self.strategy = strategy
-        self.basket = basket
-        super(BaseSavedLineFormSet, self).__init__(*args, **kwargs)
-
-    def _construct_form(self, i, **kwargs):
-        return super(BaseSavedLineFormSet, self)._construct_form(
-            i, strategy=self.strategy, basket=self.basket, **kwargs)
-
-
-SavedLineFormSet = modelformset_factory(Line, form=SavedLineForm,
-                                        formset=BaseSavedLineFormSet, extra=0,
-                                        can_delete=True)
 
 
 class BasketVoucherForm(forms.Form):
