+++ src/oscar/apps/dashboard/catalogue/views.py
@@ -17,20 +17,21 @@ from oscar.views.generic import ObjectLookupView
  ProductSearchForm,
  ProductClassForm,
  CategoryForm,
- StockRecordFormSet,
- StockAlertSearchForm,
- ProductCategoryFormSet,
- ProductImageFormSet,
- ProductRecommendationFormSet,
- ProductAttributesFormSet) \
+ StockAlertSearchForm) \
     = get_classes('dashboard.catalogue.forms',
                   ('ProductForm',
                    'ProductClassSelectForm',
                    'ProductSearchForm',
                    'ProductClassForm',
                    'CategoryForm',
-                   'StockRecordFormSet',
-                   'StockAlertSearchForm',
+                   'StockAlertSearchForm'))
+(StockRecordFormSet,
+ ProductCategoryFormSet,
+ ProductImageFormSet,
+ ProductRecommendationFormSet,
+ ProductAttributesFormSet) \
+    = get_classes('dashboard.catalogue.formsets',
+                  ('StockRecordFormSet',
                    'ProductCategoryFormSet',
                    'ProductImageFormSet',
                    'ProductRecommendationFormSet',
